Refactor log cleaning to improve file processing logic
