+++ helpers/logger.ts
@@ -101,114 +101,6 @@ export async function processLogFile(
   });
 }
 
-/**
- * Clean all raw-*.log files by removing ANSI codes and stopping after first "may be fork..." line
- */
-/**
- * Check if a large file contains a target string using streaming
- */
-async function fileContainsString(
-  filePath: string,
-  targetString: string,
-): Promise<boolean> {
-  return new Promise((resolve, reject) => {
-    const stream = fs.createReadStream(filePath, {
-      encoding: "utf8",
-      highWaterMark: 64 * 1024,
-    }); // 64KB chunks
-    let buffer = "";
-    let found = false;
-
-    stream.on("data", (chunk: string | Buffer) => {
-      const chunkStr =
-        typeof chunk === "string" ? chunk : chunk.toString("utf8");
-      buffer += chunkStr;
-
-      // Check if we found the target string
-      if (buffer.includes(targetString)) {
-        found = true;
-        stream.destroy(); // Stop reading
-        return;
-      }
-
-      // Keep only the last part of buffer to handle strings that span chunks
-      // Keep twice the target string length to be safe
-      const keepLength = targetString.length * 2;
-      if (buffer.length > keepLength) {
-        buffer = buffer.slice(-keepLength);
-      }
-    });
-
-    stream.on("end", () => {
-      resolve(found);
-    });
-
-    stream.on("error", (error) => {
-      reject(error);
-    });
-
-    stream.on("close", () => {
-      resolve(found);
-    });
-  });
-}
-
-export async function cleanAllRawLogs(): Promise<void> {
-  const logsDir = path.join(process.cwd(), "logs");
-  const outputDir = path.join(logsDir, "cleaned");
-
-  if (!fs.existsSync(logsDir)) {
-    console.log("No logs directory found");
-    return;
-  }
-
-  if (!fs.existsSync(outputDir)) {
-    await fs.promises.mkdir(outputDir, { recursive: true });
-  }
-
-  const files = await fs.promises.readdir(logsDir);
-  const rawLogFiles = files.filter(
-    (file) => file.startsWith("raw-") && file.endsWith(".log"),
-  );
-
-  if (rawLogFiles.length === 0) {
-    console.log("No raw-*.log files found to clean");
-    return;
-  }
-
-  console.log(`Found ${rawLogFiles.length} raw log files to check`);
-
-  let processedCount = 0;
-  for (const file of rawLogFiles) {
-    const inputPath = path.join(logsDir, file);
-
-    try {
-      // Check if file contains "your group may be forked" using streaming
-      const containsTargetString = await fileContainsString(inputPath, "fork");
-
-      if (!containsTargetString) {
-        console.log(
-          `Skipping ${file} - does not contain "your group may be forked"`,
-        );
-        continue;
-      }
-
-      const outputFileName = file.replace("raw-", "cleaned-");
-      const outputPath = path.join(outputDir, outputFileName);
-
-      await processLogFile(inputPath, outputPath);
-      console.log(`Cleaned: ${file} -> ${outputFileName}`);
-      processedCount++;
-    } catch (error) {
-      console.error(`Failed to process ${file}:`, error);
-    }
-  }
-
-  console.log(
-    `Processed ${processedCount} files containing "your group may be forked"`,
-  );
-}
-
 // Extend winston Logger interface for custom methods
 declare module "winston" {
   interface Logger {
